refactor: consolidate chat handlers and fix dual Query extractors
- Move all 9 chat handlers from server.rs inline to handlers/chat.rs,
  removing ~987 lines of duplicate code. The handlers/chat.rs versions
  now have full staging features (rate limiting, SSE event IDs, image
  attachments) merged with workspace scoping.
- Fix dual Query extractor bug in settings handlers: merge workspace
  field into SettingScopeQuery so axum sees one Query extractor per
  handler instead of two (which would fail at runtime).
- Restore ensure_setting_write_allowed calls lost during merge.
