+++ src/channels/web/handlers/chat.rs
@@ -1,4 +1,5 @@
-//! Chat handlers: send, approval, auth, SSE events, WebSocket, history, and shared helpers.
+//! Chat handlers: send, approval, gate resolution, auth, SSE events, WebSocket, history, threads,
+//! and shared helpers.
 
 use std::sync::Arc;
 
@@ -7,7 +8,10 @@ use crate::channels::web::auth::AuthenticatedUser;
 use crate::channels::web::handlers::workspaces::{WorkspaceQuery, resolve_requested_workspace_id};
 use crate::channels::web::server::GatewayState;
 use crate::channels::web::types::*;
-use crate::channels::web::util::{build_turns_from_db_messages, tool_error_for_display, truncate_preview};
+use crate::channels::web::util::{
+    build_turns_from_db_messages, collect_generated_images_from_tool_results,
+    enforce_generated_image_history_budget, tool_error_for_display, tool_result_preview,
+};
 use axum::{
     Json,
     extract::{Query, State, WebSocketUpgrade},
@@ -17,13 +21,22 @@ use axum::{
 use serde::Deserialize;
 use uuid::Uuid;
 
+// ── Chat send ────────────────────────────────────────────────────────────
+
 pub async fn chat_send_handler(
     State(state): State<Arc<GatewayState>>,
-    AuthenticatedUser(identity): AuthenticatedUser,
+    AuthenticatedUser(user): AuthenticatedUser,
     Query(workspace_query): Query<WorkspaceQuery>,
+    headers: axum::http::HeaderMap,
     Json(req): Json<SendMessageRequest>,
 ) -> Result<(StatusCode, Json<SendMessageResponse>), (StatusCode, String)> {
-    if !state.chat_rate_limiter.check(&identity.user_id) {
+    tracing::trace!(
+        "[chat_send_handler] Received message: content_len={}, thread_id={:?}",
+        req.content.len(),
+        req.thread_id
+    );
+
+    if !state.chat_rate_limiter.check(&user.user_id) {
         return Err((
             StatusCode::TOO_MANY_REQUESTS,
             "Rate limit exceeded. Try again shortly.".to_string(),
@@ -31,26 +44,44 @@ pub async fn chat_send_handler(
     }
 
     let workspace_id =
-        resolve_requested_workspace_id(&state, &identity, workspace_query.workspace.as_deref())
-            .await?
-            .map(|id| id.to_string());
-    let mut msg = IncomingMessage::new("gateway", &identity.user_id, &req.content);
-    if let Some(ref workspace_id) = workspace_id {
-        msg.workspace_id = Some(workspace_id.clone());
+        resolve_requested_workspace_id(&state, &user, workspace_query.workspace.as_deref()).await?;
+    let mut msg = IncomingMessage::new("gateway", &user.user_id, &req.content);
+    if let Some(workspace_id) = workspace_id {
+        msg.workspace_id = Some(workspace_id.to_string());
+    }
+    // Prefer timezone from JSON body, fall back to X-Timezone header
+    let tz = req
+        .timezone
+        .as_deref()
+        .or_else(|| headers.get("X-Timezone").and_then(|v| v.to_str().ok()));
+    if let Some(tz) = tz {
+        msg = msg.with_timezone(tz);
     }
 
-    let mut metadata = serde_json::json!({ "user_id": &identity.user_id });
-    if let Some(ref workspace_id) = workspace_id {
-        metadata["workspace_id"] = serde_json::json!(workspace_id);
+    // Always include user_id in metadata so downstream SSE broadcasts can scope events.
+    let mut meta = serde_json::json!({"user_id": &user.user_id});
+    if let Some(workspace_id) = workspace_id {
+        meta["workspace_id"] = serde_json::json!(workspace_id);
     }
     if let Some(ref thread_id) = req.thread_id {
         msg = msg.with_thread(thread_id);
-        metadata["thread_id"] = serde_json::json!(thread_id);
+        meta["thread_id"] = serde_json::json!(thread_id);
     }
-    msg = msg.with_metadata(metadata);
+    msg = msg.with_metadata(meta);
+
+    // Convert uploaded images to IncomingAttachments
+    if !req.images.is_empty() {
+        let attachments = crate::channels::web::server::images_to_attachments(&req.images);
+        msg = msg.with_attachments(attachments);
+    }
 
     let msg_id = msg.id;
-    let thread_id = msg.thread_id.clone();
+    tracing::trace!(
+        "[chat_send_handler] Created message id={}, content_len={}, images={}",
+        msg_id,
+        req.content.len(),
+        req.images.len()
+    );
 
     // Clone sender to avoid holding RwLock read guard across send().await
     let tx = {
@@ -64,6 +95,7 @@ pub async fn chat_send_handler(
             .clone()
     };
 
+    tracing::debug!("[chat_send_handler] Sending message through channel");
     tx.send(msg).await.map_err(|_| {
         (
             StatusCode::INTERNAL_SERVER_ERROR,
@@ -71,12 +103,7 @@ pub async fn chat_send_handler(
         )
     })?;
 
-    tracing::debug!(
-        message_id = %msg_id,
-        thread_id = ?thread_id,
-        content_len = req.content.len(),
-        "Message queued to agent loop"
-    );
+    tracing::debug!("[chat_send_handler] Message sent successfully, returning 202 ACCEPTED");
 
     Ok((
         StatusCode::ACCEPTED,
@@ -87,9 +114,11 @@ pub async fn chat_send_handler(
     ))
 }
 
+// ── Approval ─────────────────────────────────────────────────────────────
+
 pub async fn chat_approval_handler(
     State(state): State<Arc<GatewayState>>,
-    AuthenticatedUser(identity): AuthenticatedUser,
+    AuthenticatedUser(user): AuthenticatedUser,
     Json(req): Json<ApprovalRequest>,
 ) -> Result<(StatusCode, Json<SendMessageResponse>), (StatusCode, String)> {
     let (approved, always) = match req.action.as_str() {
@@ -125,14 +154,14 @@ pub async fn chat_approval_handler(
         )
     })?;
 
-    let mut msg = IncomingMessage::new("gateway", &identity.user_id, content);
+    let mut msg = IncomingMessage::new("gateway", &user.user_id, content);
 
     if let Some(ref thread_id) = req.thread_id {
         msg = msg.with_thread(thread_id);
         if let Some(ref sm) = state.session_manager
             && let Ok(thread_uuid) = Uuid::parse_str(thread_id)
         {
-            let session = sm.get_or_create_session(&identity.user_id).await;
+            let session = sm.get_or_create_session(&user.user_id).await;
             let sess = session.lock().await;
             if let Some(workspace_id) = sess
                 .threads
@@ -142,7 +171,7 @@ pub async fn chat_approval_handler(
             {
                 msg.workspace_id = Some(workspace_id.clone());
                 msg = msg.with_metadata(serde_json::json!({
-                    "user_id": &identity.user_id,
+                    "user_id": &user.user_id,
                     "thread_id": thread_id,
                     "workspace_id": workspace_id,
                 }));
@@ -151,7 +180,7 @@ pub async fn chat_approval_handler(
     }
 
     if msg.metadata.is_null() {
-        let mut metadata = serde_json::json!({ "user_id": &identity.user_id });
+        let mut metadata = serde_json::json!({"user_id": &user.user_id});
         if let Some(ref thread_id) = req.thread_id {
             metadata["thread_id"] = serde_json::json!(thread_id);
         }
@@ -188,7 +217,64 @@ pub async fn chat_approval_handler(
     ))
 }
 
-/// Submit an auth token directly to the extension manager, bypassing the message pipeline.
+// ── Gate resolution ──────────────────────────────────────────────────────
+
+pub async fn chat_gate_resolve_handler(
+    State(state): State<Arc<GatewayState>>,
+    AuthenticatedUser(user): AuthenticatedUser,
+    Json(req): Json<GateResolveRequest>,
+) -> Result<Json<ActionResponse>, (StatusCode, String)> {
+    match req.resolution {
+        GateResolutionPayload::Approved { always } => {
+            let action = if always { "always" } else { "approve" }.to_string();
+            let _ = chat_approval_handler(
+                State(state),
+                AuthenticatedUser(user),
+                Json(ApprovalRequest {
+                    request_id: req.request_id,
+                    action,
+                    thread_id: req.thread_id,
+                }),
+            )
+            .await?;
+            Ok(Json(ActionResponse::ok("Gate resolution accepted.")))
+        }
+        GateResolutionPayload::Denied => {
+            let _ = chat_approval_handler(
+                State(state),
+                AuthenticatedUser(user),
+                Json(ApprovalRequest {
+                    request_id: req.request_id,
+                    action: "deny".into(),
+                    thread_id: req.thread_id,
+                }),
+            )
+            .await?;
+            Ok(Json(ActionResponse::ok("Gate resolution accepted.")))
+        }
+        GateResolutionPayload::CredentialProvided { token } => {
+            let thread_id = req.thread_id.ok_or((
+                StatusCode::BAD_REQUEST,
+                "thread_id is required for credential resolution".to_string(),
+            ))?;
+            dispatch_engine_auth_resolution(&state, &user.user_id, &thread_id, token).await?;
+            Ok(Json(ActionResponse::ok("Credential submitted.")))
+        }
+        GateResolutionPayload::Cancelled => {
+            let thread_id = req.thread_id.ok_or((
+                StatusCode::BAD_REQUEST,
+                "thread_id is required for cancellation".to_string(),
+            ))?;
+            dispatch_engine_auth_resolution(&state, &user.user_id, &thread_id, "cancel".into())
+                .await?;
+            Ok(Json(ActionResponse::ok("Gate cancelled.")))
+        }
+    }
+}
+
+// ── Auth token ───────────────────────────────────────────────────────────
+
+/// Submit an auth token directly to the shared auth manager, bypassing the message pipeline.
 ///
 /// The token never touches the LLM, chat history, or SSE stream.
 pub async fn chat_auth_token_handler(
@@ -197,20 +283,58 @@ pub async fn chat_auth_token_handler(
     Query(workspace_query): Query<WorkspaceQuery>,
     Json(req): Json<AuthTokenRequest>,
 ) -> Result<Json<ActionResponse>, (StatusCode, String)> {
-    let ext_mgr = state.extension_manager.as_ref().ok_or((
-        StatusCode::SERVICE_UNAVAILABLE,
-        "Extension manager not available".to_string(),
-    ))?;
+    if let Some(ref thread_id) = req.thread_id
+        && crate::bridge::get_engine_pending_auth(&user.user_id, Some(thread_id))
+            .await
+            .is_some()
+    {
+        dispatch_engine_auth_resolution(&state, &user.user_id, thread_id, req.token.clone())
+            .await?;
+        return Ok(Json(ActionResponse::ok("Credential submitted.")));
+    }
+
+    let auth_manager = state
+        .auth_manager
+        .clone()
+        .or_else(|| {
+            state
+                .tool_registry
+                .as_ref()
+                .and_then(|tr| tr.secrets_store().cloned())
+                .or_else(|| state.secrets_store.clone())
+                .or_else(|| {
+                    state
+                        .extension_manager
+                        .as_ref()
+                        .map(|em| std::sync::Arc::clone(em.secrets()))
+                })
+                .map(|secrets| {
+                    Arc::new(crate::bridge::auth_manager::AuthManager::new(
+                        secrets,
+                        state.skill_registry.clone(),
+                        state.extension_manager.clone(),
+                        state.tool_registry.clone(),
+                    ))
+                })
+        })
+        .ok_or((
+            StatusCode::SERVICE_UNAVAILABLE,
+            "Auth manager not available".to_string(),
+        ))?;
     let auth_workspace_id =
         resolve_auth_event_workspace_scope(&state, &user, workspace_query.workspace.as_deref())
             .await?;
 
-    match ext_mgr
-        .configure_token(&req.extension_name, &req.token, &user.user_id)
+    match auth_manager
+        .submit_auth_token(&req.extension_name, &req.token, &user.user_id)
         .await
     {
         Ok(result) => {
-            let mut resp = ActionResponse::ok(result.message.clone());
+            let mut resp = if result.verification.is_some() || result.activated {
+                ActionResponse::ok(result.message.clone())
+            } else {
+                ActionResponse::fail(result.message.clone())
+            };
             resp.activated = Some(result.activated);
             resp.auth_url = result.auth_url.clone();
             resp.verification = result.verification.clone();
@@ -228,7 +352,8 @@ pub async fn chat_auth_token_handler(
                         thread_id: req.thread_id.clone(),
                     },
                 );
-            } else {
+            } else if result.activated {
+                // Clear auth mode on the active thread
                 clear_auth_mode(&state, &user.user_id).await;
 
                 state.sse.broadcast_for_user_in_workspace(
@@ -241,12 +366,25 @@ pub async fn chat_auth_token_handler(
                         thread_id: req.thread_id.clone(),
                     },
                 );
+            } else {
+                state.sse.broadcast_for_user_in_workspace(
+                    &user.user_id,
+                    auth_workspace_id.as_deref(),
+                    AppEvent::AuthCompleted {
+                        extension_name: req.extension_name.clone(),
+                        success: false,
+                        message: result.message,
+                        thread_id: req.thread_id.clone(),
+                    },
+                );
             }
 
             Ok(Json(resp))
         }
         Err(e) => {
             let msg = e.to_string();
+
+            // Re-emit auth_required for retry on validation errors
             if matches!(e, crate::extensions::ExtensionError::ValidationFailed(_)) {
                 state.sse.broadcast_for_user_in_workspace(
                     &user.user_id,
@@ -265,13 +403,26 @@ pub async fn chat_auth_token_handler(
     }
 }
 
+// ── Auth cancel ──────────────────────────────────────────────────────────
+
 /// Cancel an in-progress auth flow.
 pub async fn chat_auth_cancel_handler(
     State(state): State<Arc<GatewayState>>,
-    AuthenticatedUser(identity): AuthenticatedUser,
-    Json(_req): Json<AuthCancelRequest>,
+    AuthenticatedUser(user): AuthenticatedUser,
+    Json(req): Json<AuthCancelRequest>,
 ) -> Result<Json<ActionResponse>, (StatusCode, String)> {
-    clear_auth_mode(&state, &identity.user_id).await;
+    if let Some(ref thread_id) = req.thread_id
+        && crate::bridge::get_engine_pending_auth(&user.user_id, Some(thread_id))
+            .await
+            .is_some()
+    {
+        dispatch_engine_auth_resolution(&state, &user.user_id, thread_id, "cancel".into()).await?;
+        return Ok(Json(ActionResponse::ok("Auth cancelled")));
+    }
+
+    clear_auth_mode(&state, &user.user_id).await;
+    // Also clear engine v2 pending auth so the next message isn't consumed as a token.
+    crate::bridge::clear_engine_pending_auth(&user.user_id, req.thread_id.as_deref()).await;
     Ok(Json(ActionResponse::ok("Auth cancelled")))
 }
 
@@ -366,14 +517,39 @@ pub(crate) fn extract_last_event_id(
     })
 }
 
+/// Check whether an Origin header value points to a local address.
+///
+/// Extracts the host from the origin (handling both IPv4/hostname and IPv6
+/// literal formats) and compares it against known local addresses. Used to
+/// prevent cross-site WebSocket hijacking while allowing localhost access.
+pub(crate) fn is_local_origin(origin: &str) -> bool {
+    let host = origin
+        .strip_prefix("http://")
+        .or_else(|| origin.strip_prefix("https://"))
+        .and_then(|rest| {
+            if rest.starts_with('[') {
+                // IPv6 literal: extract "[::1]" up to and including ']'
+                rest.find(']').map(|i| &rest[..=i])
+            } else {
+                // IPv4 or hostname: take up to the first ':' (port) or '/' (path)
+                rest.split(':').next()?.split('/').next()
+            }
+        })
+        .unwrap_or("");
+
+    matches!(host, "localhost" | "127.0.0.1" | "[::1]")
+}
+
 pub async fn chat_ws_handler(
+    AuthenticatedUser(user): AuthenticatedUser,
+    Query(workspace_query): Query<WorkspaceQuery>,
     headers: axum::http::HeaderMap,
     ws: WebSocketUpgrade,
     State(state): State<Arc<GatewayState>>,
-    AuthenticatedUser(identity): AuthenticatedUser,
-    Query(workspace_query): Query<WorkspaceQuery>,
 ) -> Result<impl IntoResponse, (StatusCode, String)> {
     // Validate Origin header to prevent cross-site WebSocket hijacking.
+    // Require the header outright; browsers always send it for WS upgrades,
+    // so a missing Origin means a non-browser client trying to bypass the check.
     let origin = headers
         .get("origin")
         .and_then(|v| v.to_str().ok())
@@ -384,13 +560,7 @@ pub async fn chat_ws_handler(
             )
         })?;
 
-    let host = origin
-        .strip_prefix("http://")
-        .or_else(|| origin.strip_prefix("https://"))
-        .and_then(|rest| rest.split(':').next()?.split('/').next())
-        .unwrap_or("");
-
-    let is_local = matches!(host, "localhost" | "127.0.0.1" | "[::1]");
+    let is_local = is_local_origin(origin);
     if !is_local {
         return Err((
             StatusCode::FORBIDDEN,
@@ -398,11 +568,11 @@ pub async fn chat_ws_handler(
         ));
     }
     let workspace_id =
-        resolve_requested_workspace_id(&state, &identity, workspace_query.workspace.as_deref())
+        resolve_requested_workspace_id(&state, &user, workspace_query.workspace.as_deref())
             .await?
             .map(|id| id.to_string());
     Ok(ws.on_upgrade(move |socket| {
-        crate::channels::web::ws::handle_ws_connection(socket, state, identity, workspace_id)
+        crate::channels::web::ws::handle_ws_connection(socket, state, user, workspace_id)
     }))
 }
 
@@ -416,9 +586,97 @@ pub struct HistoryQuery {
     pub workspace: Option<String>,
 }
 
+async fn engine_pending_gate_info(
+    user_id: &str,
+    thread_id: Option<&str>,
+) -> Option<PendingGateInfo> {
+    let pending = crate::bridge::get_engine_pending_gate(user_id, thread_id)
+        .await
+        .ok()??;
+    Some(PendingGateInfo {
+        request_id: pending.request_id,
+        thread_id: pending.thread_id.to_string(),
+        gate_name: pending.gate_name,
+        tool_name: pending.tool_name,
+        description: pending.description,
+        parameters: pending.parameters,
+        resume_kind: serde_json::to_value(pending.resume_kind).unwrap_or_default(),
+    })
+}
+
+async fn history_pending_gate_info(
+    user_id: &str,
+    thread_id: Option<&str>,
+) -> Option<PendingGateInfo> {
+    if thread_id.is_some() {
+        // Thread-scoped pending gates are authoritative once the client sends a
+        // thread_id. The unscoped fallback only exists for legacy callers that
+        // do not know which thread owns the gate yet.
+        return engine_pending_gate_info(user_id, thread_id).await;
+    }
+    engine_pending_gate_info(user_id, None).await
+}
+
+async fn dispatch_engine_auth_resolution(
+    state: &GatewayState,
+    user_id: &str,
+    thread_id: &str,
+    content: String,
+) -> Result<(), (StatusCode, String)> {
+    let tx = {
+        let tx_guard = state.msg_tx.read().await;
+        tx_guard
+            .as_ref()
+            .ok_or((
+                StatusCode::SERVICE_UNAVAILABLE,
+                "Channel not started".to_string(),
+            ))?
+            .clone()
+    };
+
+    let msg = IncomingMessage::new("gateway", user_id, content).with_thread(thread_id.to_string());
+
+    tx.send(msg).await.map_err(|_| {
+        (
+            StatusCode::INTERNAL_SERVER_ERROR,
+            "Channel closed".to_string(),
+        )
+    })
+}
+
+pub(crate) fn turn_info_from_in_memory_turn(t: &crate::agent::session::Turn) -> TurnInfo {
+    TurnInfo {
+        turn_number: t.turn_number,
+        user_input: t.user_input.clone(),
+        response: t.response.clone(),
+        state: format!("{:?}", t.state),
+        started_at: t.started_at.to_rfc3339(),
+        completed_at: t.completed_at.map(|dt| dt.to_rfc3339()),
+        tool_calls: t
+            .tool_calls
+            .iter()
+            .map(|tc| ToolCallInfo {
+                name: tc.name.clone(),
+                has_result: tc.result.is_some(),
+                has_error: tc.error.is_some(),
+                result_preview: tool_result_preview(tc.result.as_ref()),
+                error: tc.error.as_deref().map(tool_error_for_display),
+                rationale: tc.rationale.clone(),
+            })
+            .collect(),
+        generated_images: collect_generated_images_from_tool_results(
+            t.turn_number,
+            t.tool_calls
+                .iter()
+                .map(|tc| (tc.tool_call_id.as_deref(), tc.result.as_ref())),
+        ),
+        narrative: t.narrative.clone(),
+    }
+}
+
 pub async fn chat_history_handler(
     State(state): State<Arc<GatewayState>>,
-    AuthenticatedUser(identity): AuthenticatedUser,
+    AuthenticatedUser(user): AuthenticatedUser,
     Query(query): Query<HistoryQuery>,
 ) -> Result<Json<HistoryResponse>, (StatusCode, String)> {
     let session_manager = state.session_manager.as_ref().ok_or((
@@ -426,9 +684,8 @@ pub async fn chat_history_handler(
         "Session manager not available".to_string(),
     ))?;
 
-    let session = session_manager
-        .get_or_create_session(&identity.user_id)
-        .await;
+    let session = session_manager.get_or_create_session(&user.user_id).await;
+    let sess = session.lock().await;
 
     let limit = query.limit.unwrap_or(50);
     let before_cursor = query
@@ -446,24 +703,27 @@ pub async fn chat_history_handler(
         })
         .transpose()?;
 
-    // Find the thread (lock only briefly to get active_thread if needed)
+    // Find the thread
     let thread_id = if let Some(ref tid) = query.thread_id {
         Uuid::parse_str(tid)
             .map_err(|_| (StatusCode::BAD_REQUEST, "Invalid thread_id".to_string()))?
     } else {
-        let sess = session.lock().await;
         sess.active_thread
             .ok_or((StatusCode::NOT_FOUND, "No active thread".to_string()))?
     };
+    let thread_id_str = thread_id.to_string();
+    let thread_scope = Some(thread_id_str.as_str());
 
     // Verify the thread belongs to the authenticated user before returning any data.
-    // This must run for both explicit thread_id and session-fallback thread_id to
-    // prevent reading arbitrary threads via the paginated DB path.
     let workspace_id =
-        resolve_requested_workspace_id(&state, &identity, query.workspace.as_deref()).await?;
-    if let Some(ref store) = state.store {
+        resolve_requested_workspace_id(&state, &user, query.workspace.as_deref()).await?;
+    // In-memory threads are already scoped by user via session_manager, but DB
+    // lookups could expose another user's conversation if the UUID is guessed.
+    if query.thread_id.is_some()
+        && let Some(ref store) = state.store
+    {
         let owned = store
-            .conversation_belongs_to_user(thread_id, &identity.user_id, workspace_id)
+            .conversation_belongs_to_user(thread_id, &user.user_id, workspace_id)
             .await
             .map_err(|e| {
                 tracing::error!(thread_id = %thread_id, error = %e, "DB error during thread ownership check");
@@ -484,74 +744,49 @@ pub async fn chat_history_handler(
             .map_err(|e| (StatusCode::INTERNAL_SERVER_ERROR, e.to_string()))?;
 
         let oldest_timestamp = messages.first().map(|m| m.created_at.to_rfc3339());
-        let turns = build_turns_from_db_messages(&messages);
+        let mut turns = build_turns_from_db_messages(&messages);
+        enforce_generated_image_history_budget(&mut turns);
         return Ok(Json(HistoryResponse {
             thread_id,
             turns,
             has_more,
             oldest_timestamp,
-            pending_gate: None,
+            pending_gate: history_pending_gate_info(&user.user_id, thread_scope).await,
         }));
     }
 
     // Try in-memory first (freshest data for active threads)
-    // Lock only when checking in-memory state
+    if let Some(thread) = sess.threads.get(&thread_id)
+        && (!thread.turns.is_empty() || thread.pending_approval.is_some())
     {
-        let sess = session.lock().await;
-        if let Some(thread) = sess.threads.get(&thread_id)
-            && (!thread.turns.is_empty() || thread.pending_approval.is_some())
-        {
-            let turns: Vec<TurnInfo> = thread
-                .turns
-                .iter()
-                .map(|t| TurnInfo {
-                    turn_number: t.turn_number,
-                    user_input: t.user_input.clone(),
-                    response: t.response.clone(),
-                    state: format!("{:?}", t.state),
-                    started_at: t.started_at.to_rfc3339(),
-                    completed_at: t.completed_at.map(|dt| dt.to_rfc3339()),
-                    tool_calls: t
-                        .tool_calls
-                        .iter()
-                        .map(|tc| ToolCallInfo {
-                            name: tc.name.clone(),
-                            has_result: tc.result.is_some(),
-                            has_error: tc.error.is_some(),
-                            result_preview: tc.result.as_ref().map(|r| {
-                                let s = match r {
-                                    serde_json::Value::String(s) => s.clone(),
-                                    other => other.to_string(),
-                                };
-                                truncate_preview(&s, 500)
-                            }),
-                            error: tc.error.as_deref().map(tool_error_for_display),
-                            rationale: tc.rationale.clone(),
-                        })
-                        .collect(),
-                    narrative: t.narrative.clone(),
-                    generated_images: Vec::new(),
-                })
-                .collect();
+        let mut turns: Vec<TurnInfo> = thread
+            .turns
+            .iter()
+            .map(turn_info_from_in_memory_turn)
+            .collect();
+        enforce_generated_image_history_budget(&mut turns);
 
-            let pending_gate = thread.pending_approval.as_ref().map(|pa| PendingGateInfo {
-                request_id: pa.request_id.to_string(),
-                thread_id: thread_id.to_string(),
-                gate_name: "approval".into(),
-                tool_name: pa.tool_name.clone(),
-                description: pa.description.clone(),
-                parameters: serde_json::to_string_pretty(&pa.parameters).unwrap_or_default(),
-                resume_kind: serde_json::json!({"Approval":{"allow_always":true}}),
+        let pending_gate = history_pending_gate_info(&user.user_id, thread_scope)
+            .await
+            .or_else(|| {
+                thread.pending_approval.as_ref().map(|pa| PendingGateInfo {
+                    request_id: pa.request_id.to_string(),
+                    thread_id: thread_id.to_string(),
+                    gate_name: "approval".into(),
+                    tool_name: pa.tool_name.clone(),
+                    description: pa.description.clone(),
+                    parameters: serde_json::to_string_pretty(&pa.parameters).unwrap_or_default(),
+                    resume_kind: serde_json::json!({"Approval":{"allow_always":true}}),
+                })
             });
 
-            return Ok(Json(HistoryResponse {
-                thread_id,
-                turns,
-                has_more: false,
-                oldest_timestamp: None,
-                pending_gate,
-            }));
-        }
+        return Ok(Json(HistoryResponse {
+            thread_id,
+            turns,
+            has_more: false,
+            oldest_timestamp: None,
+            pending_gate,
+        }));
     }
 
     // Fall back to DB for historical threads not in memory (paginated)
@@ -563,13 +798,14 @@ pub async fn chat_history_handler(
 
         if !messages.is_empty() {
             let oldest_timestamp = messages.first().map(|m| m.created_at.to_rfc3339());
-            let turns = build_turns_from_db_messages(&messages);
+            let mut turns = build_turns_from_db_messages(&messages);
+            enforce_generated_image_history_budget(&mut turns);
             return Ok(Json(HistoryResponse {
                 thread_id,
                 turns,
                 has_more,
                 oldest_timestamp,
-                pending_gate: None,
+                pending_gate: history_pending_gate_info(&user.user_id, thread_scope).await,
             }));
         }
     }
@@ -580,13 +816,13 @@ pub async fn chat_history_handler(
         turns: Vec::new(),
         has_more: false,
         oldest_timestamp: None,
-        pending_gate: None,
+        pending_gate: history_pending_gate_info(&user.user_id, thread_scope).await,
     }))
 }
 
 pub async fn chat_threads_handler(
     State(state): State<Arc<GatewayState>>,
-    AuthenticatedUser(identity): AuthenticatedUser,
+    AuthenticatedUser(user): AuthenticatedUser,
     Query(workspace_query): Query<WorkspaceQuery>,
 ) -> Result<Json<ThreadListResponse>, (StatusCode, String)> {
     let session_manager = state.session_manager.as_ref().ok_or((
@@ -594,77 +830,74 @@ pub async fn chat_threads_handler(
         "Session manager not available".to_string(),
     ))?;
 
-    let session = session_manager
-        .get_or_create_session(&identity.user_id)
-        .await;
+    let session = session_manager.get_or_create_session(&user.user_id).await;
+    let sess = session.lock().await;
 
     // Try DB first for persistent thread list
     if let Some(ref store) = state.store {
         let workspace_id =
-            resolve_requested_workspace_id(&state, &identity, workspace_query.workspace.as_deref())
+            resolve_requested_workspace_id(&state, &user, workspace_query.workspace.as_deref())
                 .await?;
         // Auto-create assistant thread if it doesn't exist
         let assistant_id = store
-            .get_or_create_assistant_conversation(&identity.user_id, workspace_id, "gateway")
+            .get_or_create_assistant_conversation(&user.user_id, workspace_id, "gateway")
             .await
             .map_err(|e| (StatusCode::INTERNAL_SERVER_ERROR, e.to_string()))?;
 
-        if let Ok(summaries) = store
-            .list_conversations_all_channels(&identity.user_id, workspace_id, 50)
+        match store
+            .list_conversations_all_channels(&user.user_id, workspace_id, 50)
             .await
         {
-            let mut assistant_thread = None;
-            let mut threads = Vec::new();
+            Ok(summaries) => {
+                let mut assistant_thread = None;
+                let mut threads = Vec::new();
 
-            for s in &summaries {
-                let info = ThreadInfo {
-                    id: s.id,
-                    state: "Idle".to_string(),
-                    turn_count: s.message_count.max(0) as usize,
-                    created_at: s.started_at.to_rfc3339(),
-                    updated_at: s.last_activity.to_rfc3339(),
-                    title: s.title.clone(),
-                    thread_type: s.thread_type.clone(),
-                    channel: Some(s.channel.clone()),
-                };
+                for s in &summaries {
+                    let info = ThreadInfo {
+                        id: s.id,
+                        state: "Idle".to_string(),
+                        turn_count: s.message_count.max(0) as usize,
+                        created_at: s.started_at.to_rfc3339(),
+                        updated_at: s.last_activity.to_rfc3339(),
+                        title: s.title.clone(),
+                        thread_type: s.thread_type.clone(),
+                        channel: Some(s.channel.clone()),
+                    };
 
-                if s.id == assistant_id {
-                    assistant_thread = Some(info);
-                } else {
-                    threads.push(info);
+                    if s.id == assistant_id {
+                        assistant_thread = Some(info);
+                    } else {
+                        threads.push(info);
+                    }
                 }
+
+                // If assistant wasn't in the list (0 messages), synthesize it
+                if assistant_thread.is_none() {
+                    assistant_thread = Some(ThreadInfo {
+                        id: assistant_id,
+                        state: "Idle".to_string(),
+                        turn_count: 0,
+                        created_at: chrono::Utc::now().to_rfc3339(),
+                        updated_at: chrono::Utc::now().to_rfc3339(),
+                        title: None,
+                        thread_type: Some("assistant".to_string()),
+                        channel: Some("gateway".to_string()),
+                    });
+                }
+
+                return Ok(Json(ThreadListResponse {
+                    assistant_thread,
+                    threads,
+                    active_thread: sess.active_thread,
+                }));
             }
-
-            // If assistant wasn't in the list (0 messages), synthesize it
-            if assistant_thread.is_none() {
-                assistant_thread = Some(ThreadInfo {
-                    id: assistant_id,
-                    state: "Idle".to_string(),
-                    turn_count: 0,
-                    created_at: chrono::Utc::now().to_rfc3339(),
-                    updated_at: chrono::Utc::now().to_rfc3339(),
-                    title: None,
-                    thread_type: Some("assistant".to_string()),
-                    channel: Some("gateway".to_string()),
-                });
+            Err(e) => {
+                tracing::error!(user_id = %user.user_id, error = %e, "DB error listing threads; falling back to in-memory");
             }
-
-            // Read active thread while holding minimal lock (just before return)
-            let active_thread = {
-                let sess = session.lock().await;
-                sess.active_thread
-            };
-
-            return Ok(Json(ThreadListResponse {
-                assistant_thread,
-                threads,
-                active_thread,
-            }));
         }
     }
 
     // Fallback: in-memory only (no assistant thread without DB)
-    let sess = session.lock().await;
     let mut sorted_threads: Vec<_> = sess.threads.values().collect();
     sorted_threads.sort_by_key(|t| std::cmp::Reverse(t.updated_at));
     let threads: Vec<ThreadInfo> = sorted_threads
@@ -681,19 +914,16 @@ pub async fn chat_threads_handler(
         })
         .collect();
 
-    let active_thread = sess.active_thread;
-    drop(sess); // Explicit drop to release lock
-
     Ok(Json(ThreadListResponse {
         assistant_thread: None,
         threads,
-        active_thread,
+        active_thread: sess.active_thread,
     }))
 }
 
 pub async fn chat_new_thread_handler(
     State(state): State<Arc<GatewayState>>,
-    AuthenticatedUser(identity): AuthenticatedUser,
+    AuthenticatedUser(user): AuthenticatedUser,
     Query(workspace_query): Query<WorkspaceQuery>,
 ) -> Result<Json<ThreadInfo>, (StatusCode, String)> {
     let session_manager = state.session_manager.as_ref().ok_or((
@@ -701,12 +931,10 @@ pub async fn chat_new_thread_handler(
         "Session manager not available".to_string(),
     ))?;
 
-    let session = session_manager
-        .get_or_create_session(&identity.user_id)
-        .await;
+    let session = session_manager.get_or_create_session(&user.user_id).await;
     let (thread_id, info) = {
         let mut sess = session.lock().await;
-        let thread = sess.create_thread(Some("web"));
+        let thread = sess.create_thread(Some("gateway"));
         let id = thread.id;
         let info = ThreadInfo {
             id: thread.id,
@@ -725,13 +953,13 @@ pub async fn chat_new_thread_handler(
     // so that the subsequent loadThreads() call from the frontend sees it.
     if let Some(ref store) = state.store {
         let workspace_id =
-            resolve_requested_workspace_id(&state, &identity, workspace_query.workspace.as_deref())
+            resolve_requested_workspace_id(&state, &user, workspace_query.workspace.as_deref())
                 .await?;
         match store
             .ensure_conversation(
                 thread_id,
                 "gateway",
-                &identity.user_id,
+                &user.user_id,
                 workspace_id,
                 None,
                 Some("gateway"),
@@ -740,7 +968,7 @@ pub async fn chat_new_thread_handler(
         {
             Ok(true) => {}
             Ok(false) => tracing::warn!(
-                user = %identity.user_id,
+                user = %user.user_id,
                 thread_id = %thread_id,
                 "Skipped persisting new thread due to ownership/channel conflict"
             ),
@@ -760,8 +988,8 @@ pub async fn chat_new_thread_handler(
 
 #[cfg(test)]
 mod tests {
+    use super::*;
     use crate::channels::web::util::build_turns_from_db_messages;
-    use uuid::Uuid;
 
     #[test]
     fn test_build_turns_from_db_messages_complete() {
@@ -932,5 +1160,58 @@ mod tests {
         assert!(turns[0].tool_calls.is_empty());
         assert_eq!(turns[0].response.as_deref(), Some("Hi!"));
         assert_eq!(turns[0].state, "Completed");
+    }
+
+    #[test]
+    fn test_is_local_origin_localhost() {
+        assert!(is_local_origin("http://localhost:3001"));
+        assert!(is_local_origin("http://localhost"));
+        assert!(is_local_origin("https://localhost:3001"));
+    }
+
+    #[test]
+    fn test_is_local_origin_ipv4() {
+        assert!(is_local_origin("http://127.0.0.1:3001"));
+        assert!(is_local_origin("http://127.0.0.1"));
+    }
+
+    #[test]
+    fn test_is_local_origin_ipv6() {
+        assert!(is_local_origin("http://[::1]:3001"));
+        assert!(is_local_origin("http://[::1]"));
+    }
+
+    #[test]
+    fn test_is_local_origin_rejects_remote() {
+        assert!(!is_local_origin("http://evil.com"));
+        assert!(!is_local_origin("http://localhost.evil.com"));
+        assert!(!is_local_origin("http://192.168.1.1:3001"));
+    }
+
+    #[test]
+    fn test_is_local_origin_rejects_garbage() {
+        assert!(!is_local_origin("not-a-url"));
+        assert!(!is_local_origin(""));
+    }
+
+    #[test]
+    fn test_in_memory_turn_info_unwraps_wrapped_tool_error_for_display() {
+        let mut thread = crate::agent::session::Thread::new(Uuid::new_v4(), Some("gateway"));
+        thread.start_turn("Fetch example");
+        {
+            let turn = thread.turns.last_mut().expect("turn");
+            turn.record_tool_call("http", serde_json::json!({"url": "https://example.com"}));
+            turn.record_tool_error(
+                "<tool_output name=\"http\">\nTool 'http' failed: timeout\n</tool_output>",
+            );
+        }
+
+        let info = turn_info_from_in_memory_turn(&thread.turns[0]);
+
+        assert_eq!(info.tool_calls.len(), 1);
+        assert_eq!(
+            info.tool_calls[0].error.as_deref(),
+            Some("Tool 'http' failed: timeout")
+        );
     }
 }
